+++ x-pack/plugins/transform/public/app/sections/create_transform/components/step_details/step_details_form.tsx
@@ -29,7 +29,7 @@ import { toMountPoint } from '@kbn/react-kibana-mount';
 import { CreateDataViewForm } from '@kbn/ml-data-view-utils/components/create_data_view_form_row';
 import { DestinationIndexForm } from '@kbn/ml-creation-wizard-utils/components/destination_index_form';
 
-import { retentionPolicyMaxAgeInvalidErrorMessage } from '../../../../common/constants/validation_messages';
+import { retentionPolicyMaxAgeInvalidErrorMessage } from '../../../../common/validators/messages';
 import { DEFAULT_TRANSFORM_FREQUENCY } from '../../../../../../common/constants';
 import { TransformId } from '../../../../../../common/types/transform';
 import { isValidIndexName } from '../../../../../../common/utils/es_utils';
@@ -54,10 +54,10 @@ import {
 } from '../../../../common';
 import { EsIndexName } from './common';
 import {
-  continuousModeDelayValidator,
+  isContinuousModeDelay,
+  isRetentionPolicyMaxAge,
+  isTransformWizardFrequency,
   integerRangeMinus1To100Validator,
-  retentionPolicyMaxAgeValidator,
-  transformFrequencyValidator,
   transformSettingsPageSearchSizeValidator,
 } from '../../../../common/validators';
 import { StepDefineExposedState } from '../step_define/common';
@@ -260,7 +260,7 @@ export const StepDetailsForm: FC<StepDetailsFormProps> = React.memo(
       isContinuousModeAvailable ? sourceIndexDateFieldNames[0] : ''
     );
     const [continuousModeDelay, setContinuousModeDelay] = useState(defaults.continuousModeDelay);
-    const isContinuousModeDelayValid = continuousModeDelayValidator(continuousModeDelay);
+    const isContinuousModeDelayValid = isContinuousModeDelay(continuousModeDelay);
 
     // Retention Policy
     const isRetentionPolicyAvailable = destIndexAvailableTimeFields.length > 0;
@@ -274,7 +274,7 @@ export const StepDetailsForm: FC<StepDetailsFormProps> = React.memo(
       defaults.retentionPolicyMaxAge
     );
     const retentionPolicyMaxAgeEmpty = retentionPolicyMaxAge === '';
-    const isRetentionPolicyMaxAgeValid = retentionPolicyMaxAgeValidator(retentionPolicyMaxAge);
+    const isRetentionPolicyMaxAgeValid = isRetentionPolicyMaxAge(retentionPolicyMaxAge);
 
     useEffect(() => {
       // Reset retention policy settings when the user disables the whole option
@@ -308,7 +308,7 @@ export const StepDetailsForm: FC<StepDetailsFormProps> = React.memo(
     const dataViewTitleExists = dataViewTitles?.some((name) => destinationIndex === name) ?? false;
 
     const [transformFrequency, setTransformFrequency] = useState(defaults.transformFrequency);
-    const isTransformFrequencyValid = transformFrequencyValidator(transformFrequency);
+    const isTransformFrequencyValid = isTransformWizardFrequency(transformFrequency);
 
     const [transformSettingsMaxPageSearchSize, setTransformSettingsMaxPageSearchSize] = useState<
       number | undefined
