+++ x-pack/plugins/transform/public/app/sections/transform_management/components/edit_transform_flyout/use_edit_transform_flyout.tsx
@@ -10,12 +10,9 @@ import React, { useContext, useRef, type FC } from 'react';
 import { createStore, useStore } from 'zustand';
 import { produce } from 'immer';
 
-import { i18n } from '@kbn/i18n';
-import { numberValidator } from '@kbn/ml-agg-utils';
 import { getNestedProperty, setNestedProperty } from '@kbn/ml-nested-property';
 
 import { createContext } from 'react';
-import { retentionPolicyMaxAgeInvalidErrorMessage } from '../../../../common/constants/validation_messages';
 import { PostTransformsUpdateRequestSchema } from '../../../../../../common/api_schemas/update_transforms';
 import {
   DEFAULT_TRANSFORM_FREQUENCY,
@@ -23,13 +20,20 @@ import {
 } from '../../../../../../common/constants';
 import { TransformConfigUnion } from '../../../../../../common/types/transform';
 
+// Note on the form validation and input components used:
+// All inputs use `EuiFieldText` which means all form values will be treated as strings.
+// This means we cast other formats like numbers coming from the transform config to strings,
+// then revalidate them and cast them again to number before submitting a transform update.
+// We do this so we have fine grained control over field validation and the option to
+// cast to special values like `null` for disabling `docs_per_second`.
 import {
-  isValidFrequency,
-  isValidRetentionPolicyMaxAge,
-  ParsedDuration,
+  frequencyValidator,
+  integerAboveZeroValidator,
   transformSettingsNumberOfRetriesValidator,
   transformSettingsPageSearchSizeValidator,
-  Validator,
+  retentionPolicyMaxAgeValidator,
+  stringValidator,
+  type Validator,
 } from '../../../../common/validators';
 
 // This custom hook uses `zustand` to provide a generic framework to manage form state
@@ -63,7 +67,7 @@ export interface FormField {
   isNullable: boolean;
   isOptional: boolean;
   section?: EditTransformFormSections;
-  validator: keyof typeof validate;
+  validator: Validator;
   value: string;
   valueParser: (value: string) => any;
 }
@@ -99,108 +103,6 @@ type EditTransformFlyoutSectionsState = Record<EditTransformFormSections, FormSe
 //   For example, if the `pipeline` field was changed, it's necessary to make the `index`
 //   field part of the request, otherwise the update would fail.
 
-// Note on the form validation and input components used:
-// All inputs use `EuiFieldText` which means all form values will be treated as strings.
-// This means we cast other formats like numbers coming from the transform config to strings,
-// then revalidate them and cast them again to number before submitting a transform update.
-// We do this so we have fine grained control over field validation and the option to
-// cast to special values like `null` for disabling `docs_per_second`.
-const numberAboveZeroNotValidErrorMessage = i18n.translate(
-  'xpack.transform.transformList.editFlyoutFormNumberAboveZeroNotValidErrorMessage',
-  {
-    defaultMessage: 'Value needs to be an integer above zero.',
-  }
-);
-
-export const integerAboveZeroValidator: Validator = (value) =>
-  !(value + '').includes('.') && numberValidator({ min: 1, integerOnly: true })(+value) === null
-    ? []
-    : [numberAboveZeroNotValidErrorMessage];
-
-const requiredErrorMessage = i18n.translate(
-  'xpack.transform.transformList.editFlyoutFormRequiredErrorMessage',
-  {
-    defaultMessage: 'Required field.',
-  }
-);
-const stringNotValidErrorMessage = i18n.translate(
-  'xpack.transform.transformList.editFlyoutFormStringNotValidErrorMessage',
-  {
-    defaultMessage: 'Value needs to be of type string.',
-  }
-);
-export const stringValidator: Validator = (value, isOptional = true) => {
-  if (typeof value !== 'string') {
-    return [stringNotValidErrorMessage];
-  }
-
-  if (value.length === 0 && !isOptional) {
-    return [requiredErrorMessage];
-  }
-
-  return [];
-};
-
-function parseDurationAboveZero(arg: unknown, errorMessage: string): ParsedDuration | string[] {
-  if (typeof arg !== 'string' || arg === null) {
-    return [stringNotValidErrorMessage];
-  }
-
-  // split string by groups of numbers and letters
-  const regexStr = arg.match(/[a-z]+|[^a-z]+/gi);
-
-  // only valid if one group of numbers and one group of letters
-  if (regexStr === null || (Array.isArray(regexStr) && regexStr.length !== 2)) {
-    return [frequencyNotValidErrorMessage];
-  }
-
-  const number = +regexStr[0];
-  const timeUnit = regexStr[1];
-
-  // only valid if number is an integer above 0
-  if (isNaN(number) || !Number.isInteger(number) || number === 0) {
-    return [frequencyNotValidErrorMessage];
-  }
-
-  return { number, timeUnit };
-}
-
-// Only allow frequencies in the form of 1s/1h etc.
-const frequencyNotValidErrorMessage = i18n.translate(
-  'xpack.transform.transformList.editFlyoutFormFrequencyNotValidErrorMessage',
-  {
-    defaultMessage: 'The frequency value is not valid.',
-  }
-);
-export const frequencyValidator: Validator = (arg) => {
-  const parsedArg = parseDurationAboveZero(arg, frequencyNotValidErrorMessage);
-
-  if (Array.isArray(parsedArg)) {
-    return parsedArg;
-  }
-
-  return isValidFrequency(parsedArg) ? [] : [frequencyNotValidErrorMessage];
-};
-
-export const retentionPolicyMaxAgeValidator: Validator = (arg) => {
-  const parsedArg = parseDurationAboveZero(arg, retentionPolicyMaxAgeInvalidErrorMessage);
-
-  if (Array.isArray(parsedArg)) {
-    return parsedArg;
-  }
-
-  return isValidRetentionPolicyMaxAge(parsedArg) ? [] : [retentionPolicyMaxAgeInvalidErrorMessage];
-};
-
-const validate = {
-  string: stringValidator,
-  frequency: frequencyValidator,
-  integerAboveZero: integerAboveZeroValidator,
-  transformSettingsNumberOfRetriesValidator,
-  transformSettingsPageSearchSizeValidator,
-  retentionPolicyMaxAgeValidator,
-} as const;
-
 export const initializeFormField = (
   formFieldName: EditTransformFormFields,
   configFieldName: string,
@@ -219,7 +121,7 @@ export const initializeFormField = (
     errorMessages: [],
     isNullable: false,
     isOptional: true,
-    validator: 'string',
+    validator: stringValidator,
     value,
     valueParser: (v) => v,
     ...(overloads !== undefined ? { ...overloads } : {}),
@@ -331,7 +233,7 @@ export const getDefaultState = (config: TransformConfigUnion): EditTransformFlyo
     description: initializeFormField('description', 'description', config),
     frequency: initializeFormField('frequency', 'frequency', config, {
       defaultValue: DEFAULT_TRANSFORM_FREQUENCY,
-      validator: 'frequency',
+      validator: frequencyValidator,
     }),
 
     // dest.*
@@ -353,7 +255,7 @@ export const getDefaultState = (config: TransformConfigUnion): EditTransformFlyo
     docsPerSecond: initializeFormField('docsPerSecond', 'settings.docs_per_second', config, {
       isNullable: true,
       isOptional: true,
-      validator: 'integerAboveZero',
+      validator: integerAboveZeroValidator,
       valueParser: (v) => (v === '' ? null : +v),
     }),
     maxPageSearchSize: initializeFormField(
@@ -364,7 +266,7 @@ export const getDefaultState = (config: TransformConfigUnion): EditTransformFlyo
         defaultValue: `${DEFAULT_TRANSFORM_SETTINGS_MAX_PAGE_SEARCH_SIZE}`,
         isNullable: true,
         isOptional: true,
-        validator: 'transformSettingsPageSearchSizeValidator',
+        validator: transformSettingsPageSearchSizeValidator,
         valueParser: (v) => +v,
       }
     ),
@@ -376,7 +278,7 @@ export const getDefaultState = (config: TransformConfigUnion): EditTransformFlyo
         defaultValue: undefined,
         isNullable: true,
         isOptional: true,
-        validator: 'transformSettingsNumberOfRetriesValidator',
+        validator: transformSettingsNumberOfRetriesValidator,
         valueParser: (v) => +v,
       }
     ),
@@ -402,7 +304,7 @@ export const getDefaultState = (config: TransformConfigUnion): EditTransformFlyo
         isNullable: false,
         isOptional: true,
         section: 'retentionPolicy',
-        validator: 'retentionPolicyMaxAgeValidator',
+        validator: retentionPolicyMaxAgeValidator,
       }
     ),
   },
@@ -428,7 +330,7 @@ const updateFormField = (state: FormField, value: string): FormField =>
     d.errorMessages =
       state.isOptional && typeof value === 'string' && value.length === 0
         ? []
-        : validate[state.validator](value, state.isOptional);
+        : state.validator(value, state.isOptional);
     d.value = value;
   });
